Avoid performance issues when inspecting repo
By default, `skopeo inspect` will fetch a list of all the tags in the
repository. This can be an expensive operation if the repository has a
large amount of tags. A registry may paginate the results causing skopeo
to trigger an unbounded amount of queries to the registry.

Use the `--no-tags` parameter to avoid fetching the tags altogether
since they're not used.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>