cosign verify and verify-attestation Tekton tasks
This adds two tasks that take an `IMAGE` and a `PUBLIC_KEY` parameters
and verify the signature and attestation using the `cosign` binary.

I wanted the minimum of a shell script wrapper around `cosign`, so I
ended up not reusing the scripts in `appstudio-utils/util-scripts`.

The only clever bit in this is that the `PUBLIC_KEY` can be specified as
a PEM-encoded (BASE64 with header/footer) or as a valid path/URI
supported by `cosign`. This is to facilitate providing the `PUBLIC_KEY`
by means not directly supported by `cosign`, for example via an
environment variable.

In addition to that unit tests using shellspec[1] are added. To run,
install shellspec locally and run `$ shellspec` from the repository
root.

Ref. https://issues.redhat.com/browse/HACBS-106

[1] https://shellspec.info/