Add opa cli tool to appstudio-utils docker image
OPA is the tool used for for validating rego policies such as the
ones that will be used by the Enterprise Contract to validate
pipeline runs.

It's not clear yet whether the final version of the Enterprise
Contract tasks will use this base image, but there's a reasonable
chance they will, and either way it seems like this is generally
useful tool to add alongside the others.

See also:
* https://github.com/open-policy-agent/opa/releases
* https://www.openpolicyagent.org/
* https://issues.redhat.com/browse/HACBS-233
* https://issues.redhat.com/browse/HACBS-235