Fetch policy configuration from ConfigMap
For the policy configuration, namely specifying what rules are
non-blocking we can now provide a ConfigMap named `ec-policy` in the
namespace the task is running in. From the ConfigMap a key with the name
`policy.json` needs to hold the JSON file with the policy configuration.
If the ConfigMap doesn't exist the current fixed value is used as a
fallback.
This will most likely be replaced with a EnterpriseContractPolicy custom
resource in the future.

Ref. https://issues.redhat.com/browse/HACBS-244