Make sure error is displayed
A file clash was detected when executing task and this cause was printed
on the output:

```
[fetch-test-data] /appstudio-utils/util-scripts/fetch-test-data.sh: line 24: $( json-data-file test ${task_name} ): ambiguous redirect
```

This is caused by `json-data-file` returning a string with whitespace,
the error message. And `fetch-test-data.sh` not quoting the result of
that evaluation.

This adds quoting and also makes sure that in the `json-data-file` the
error is printed to the standard error, otherwise it'll be captured and
not displayed making troubleshooting difficult.