contract: Bash function for jq data merges
We're aiming to switch to conftest instead of opa for running the
contract policy rules. Unlike opa, which applies the directory
structure to the data, conftest flattens everything into one object,
which means the layout we're using now in $DATA_DIR causes merge
errors for conftest.

This bash function is intended to help avoid conftest merge errors
by applying the required data structure manually in a single file.