Produce materials for the build pipelines
When Tekton Chains attests a PipelineRun, it looks for the presence of
either results, or parameters, named CHAINS-GIT_URL and
CHAINS-GIT_COMMIT. If found, this information is then propagated to the
.predicate.materials section of the SLSA Provenance attestation.

This commit makes the build pipelines produce this information in their
results. Although using parameters is also a possibility, it has a
stronger impact on the pipeline usability since users would have to
specify a different parameter for the git url and commit.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>