Remove `cosign-verify-*` tasks
The problem is with the `cosign-verify-attestation` task that attempts
to store the attestation JSON into `VERIFY_ATTESTATION_JSON` Tekton
result. Given that today we're creating PipelineRun attestations, the
size of which is >4096 (around 20KB for a buildah build), we want to
stop using these. The `verify-enterprise-contract` (v1) task will run
`cosign` to validate and fetch the attestation and the attestation JSON
will be stored on a shared volume, so we don't have this issue there.