Fix passing of `STRICT` parameter to VEC V2
For `verify-enterprise-task` we need to pass `"1"` as `STRICT_POLICY`
PipelineRun parameter, and for `verify-enterprise-task-v2` we need to
pass `"true"` as `STRICT` PipelineRun parameter, for the enforcement to
be strict.

Likewise we can pass anything but `"1"` to as `STRICT_POLICY`, and
almost anything but `"true"` for `STRICT` parameter for the enforcement
not to be strict.

This makes it difficult to have a single PipelineRun parameter, so this
introduces a second parameter to cater to the V2 task.

Ref. https://issues.redhat.com/browse/HACBS-762