Set HOME for verify-ec2 task
As recommended by https://access.redhat.com/solutions/6956010

This allows running the verify-enterprise-contract-v2 task with the
"restricted" scc (SecurityContextConstraint). This is the scc assigned
to ServiceAccounts by default, inluding the "default" ServiceAccount.

Since the release pipeline executes with a custom ServiceAccount, this
makes it easier to use the task in that environment.

https://issues.redhat.com/browse/HACBS-880

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>