Delete verify-pipeline.sh helper script
The script hack/verify-pipeline.sh was used to verify a pipeline
definition against Enterprise Contract policies. The script used a
combination of bash scripts to achieve this.

Support for verifying a pipeline definition has been added to the
ec-cli[0]. This is the preferred mechanism to perform this action.

Instead of using:
    ./hack/verify-pipeline.sh my-pipeline.yaml

Use now:
    ec validate pipeline --pipeline-file my-pipeline.yaml

Since the verify-pipeline.sh script was the only entity using the
various bash script utilities, those have also been removed.

https://issues.redhat.com/browse/HACBS-536

[0] https://github.com/hacbs-contract/ec-cli

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>