PLNSRVCE-384: apply network isolation label to build task
Motivation:

The network isolation label is only be applied to build task running in
HACBS pipeline, but the generated tasks are shared by both appstudio and
HACBS originally. So, the tasks and pipelines should be handled for
appstudio and HACBS individually.

Note that the network isolation label is not applied to Java build at
current stage.

Major changes:

* build-and-push.sh is refactored.
* Tasks under tasks/ are grouped into base/ and hacbs/ directories. base
  directory holds the all original tasks, from which to generate
  appstudio task. And, hacbs directory contains patches to extend for
  HACBS specifically.
* A patch is created under tasks/hacbs/ to apply the network isolation
  label inside buildah task.
* A new bundle image repo hacbs-tasks is introduced for HACBS task
  bundle images.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>