Tag newest pipeline bundle images with `devel`
For ease of consumption in development folk would like to utilize the
newest versions of the pipeline bundle images with a fixed tag: `devel`.

This re-taggs the pipeline bundle images with the `devel` if
`SKIP_DEVEL_TAG` environment variable is not empty and image push is
directed to `redhat-appstudio` organization.

Fixes https://issues.redhat.com/browse/HACBS-1088