Make sure `hermetic-build` result is produced
If the Tekton Task result is not produced any when expression using that
result will evaluate to `false`. The `hermetic-build` result of the
`git-clone` Task would be set to `true` if and only if the
`cachi2.params` exists in the workspace, if the file does not exist it
will not be set.
The current default is that a when expression evaluating to false skips
the task where the when expression is defined, but also all tasks that
`runAfter` it. See[1],`scope-when-expressions-to-task` feature flag[2]
and TEP-0059[3].
Without the `cachi2.params` file in the workspace, the when condition on
the `prefetch-dependencies` would evaluate to false, and the
`build-container` step that has `runAfter: prefetch-dependencies`
would be skipped as well, cascading to every other step in the pipeline
other than `show-summary` and `sast-snyk-check` tasks.
With this change the `hermetic-build` result is always set to either
`"true"` or `"false"`.

[1] https://tekton.dev/docs/pipelines/pipelines/#guard-task-execution-using-when-expressions
[2] https://tekton.dev/docs/operator/tektonpipeline/
[3] https://github.com/tektoncd/community/blob/main/teps/0059-skipping-strategies.md