Use resource dependencies
The parameters git and image URLs should be taken from the
`clone-repository` and `build-container` tasks not from the user input.

This creates a resource dependency between the `sanity-inspect-image`
task and the `build-container` task, so if the `build-container` task is
skipped the `sanity-inspect-image` task will not be run. Not creating a
resource dependency leads to `sanity-inspect-image` task running with an
image reference pointing to an image that could not have been built in
the pipeline when the `build-container` task is skipped.

Similarly for the `show-summary` task, as it now depends on
`clone-repository` and `build-container` tasks running. And in addition
to that the `show-summary` now includes git revision that was cloned,
and since the image SHA-256 digest is passed the exact image that was
built.

[1] https://tekton.dev/docs/pipelines/pipelines/#guard-task-execution-using-when-expressions