Use digest if image ref has tag and digest
The `sanity-inspect-image` Task is using `skopeo` that currently doesn't
support image references containing both tag and digest
(see https://github.com/containers/image/issues/1736). The task has been
modified to check if the image reference contains both and in that case
the tag is removed and digest is used solely to reference the image.