Remove trailing newline from BASE_IMAGE* results
In `sanity-inspect-image` the `BASE_IMAGE` and `BASE_IMAGE_REPOSITORY`
result is set and later on used in the `deprecated-image-check`.

Currently it contains a trailing newline. The trailing newline ends up
being passed literally to `curl` used in the `deprecated-image-check` as
the URL is quoted there. This leads to curl failing with status `000`.

The `deprecated-image-check` Task doesn't fail, but the
`HACBS_TEST_OUTPUT` results in an error.

No matter how hard I tried see where the issue was and where the newline
was emitted from for the `BASE_IMAGE_REPOSITORY` result, i.e. was it
from jq, cut or tee, it seemed to originate from none of those. That's
why I ended up rewriting that in a single jq invocation and analogous to
`BASE_IMAGE` passed via variable and echo to tee. Seems to have solved
the issue.