Run the show-summary task always
This removes the resource dependency on the `build-container` task from
the `show-summary` task. This way the `show-summary` task will be run
regardless of the `build-container` task running, i.e. always.

Note that the resource dependency on the `clone-repository` task
remains. Since `clone-repository` task will not be skipped, like the
`build-container` task might, this is not an issue.