Add pipeline parameters to set hermeticity and content prefetching
The hermetic parameter will determine if the build should be executed
with network isolation.

The prefetch-input parameter is used to configure Cachi2 to properly
prefetch the dependencies needed by the build task. If this parameter
is not empty, it also means that the Cachi2 output directories should
be mounted during the image build.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>