Run the git-clone task as root
Currently, other tasks in the pipeline assume that the source was cloned
as root. Make sure the step will indeed run as root even if the default
USER in the git-init image is non-root.

Otherwise, for example the prefetch-dependencies task (which runs as
root) may fail to fetch git tags when processing Go modules:

    cmdline: git fetch -v --force --tags origin
    stderr: 'fatal: detected dubious ownership in repository at '/workspace/source''

Signed-off-by: Adam Cmiel <acmiel@redhat.com>