Create variations of the ec pipeline
Copies of the enterprise-contract.yaml are now available with a
different default value for the POLICY_CONFIGURATION parameter. This is
a workaround until the UI can provide sufficient controls to allow users
to provide parameters.

We cannot use kustomize to make these variations because these pipelines
will be used with Tekton's git resolver. This requires the pipeline
definition to appear in full in git.

https://issues.redhat.com/browse/HACBS-2294

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>