Add check for symlinks pointing outside of repo
If the check is enabled and the cloned repository contains symlinks
pointing outside of the repo, the build will fail.

This is implemented as separate step so different image with the
necessary tools can be used. It also uses bash instead of sh, so
that we can use process substition. This is just so the value of
FOUND_SYMLINK_POINTING_OUTSIDE_OF_REPO can be changed in the pipeline
command, otherwise the that part of the pipeline would run in a subshell
and the value of that variable would be lost.

STONEBLD-1328

Signed-off-by: mkosiarc <mkosiarc@redhat.com>