Create OPA data bundle of acceptable tkn bundles
EC has the concept of acceptable tekton bundles. These are OCI image
references that specify which tekton bundles are acceptable to build
artifacts in RHTAP. Today, these are tracked in git. The push.yaml
pipeline of this repo is responsible for making a pull request in the
relevant EC repo update the list. However, this is subject to timing
issues when the pull request is not merged fast enough, causing some of
the references to be lost.

This commit introduces a new approach by tracking the list of acceptable
tekton bundles in an OPA data bundle, which is just another type of OCI
image. By doing this, the changes are deployed immediately,
significantly lowering the risk of timing issues.

For now, the list is maintained in two places. This facilitate the
transition. See the epic https://issues.redhat.com/browse/HACBS-2554 for
more details.

Tracker for this change:
https://issues.redhat.com/browse/HACBS-1796

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>