Emit SBOM_BLOB_URL result on build tasks
This changes the buildah, s2i-java, and s2i-nodejs tasks to emit a new
task result, SBOM_BLOB_URL. This result includes an image reference by
digest to the SBOM of the image.

By doing so, we establish a strong link between the image and the SBOM
via the SLSA Provenance.

https://issues.redhat.com/browse/EC-175

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>