WIP: DIY Trusted Artifacts
This PR modifies the git-clone and the buildah task to use a pattern
that ensures the buildah task is using the unmodified contents of the
git repo cloned by the git-clone task.

There is an ongoing effort in the Tekton community to introduce the
concept of trusted artifacts, TEP-0130
https://github.com/tektoncd/community/pull/1044

This change is directly influenced by that effort.

The approach taken here is a stop gap until the platform, i.e. Tekton,
supports the idea of trusted artifacts natively.

Here's how it works...

I created a couple of little bash scripts to encapsulate producing and
consuming an artifact: https://github.com/lcarva/trusted-artifacts
These are available in the image: https://quay.io/lucarval/trusted-artifacts-demo
This is meant to make it easier to use them in a Tekton task, instead of
just stuffing bash scripts all over the place.

The creation of an artifact is pretty simple. It just creates a tarball
containing whichever files are deemed part of the artifact. In this
case, the artifact is the directory containing a cloned git repo.

The consumption of the artifact is the inverse. It extracts the tarball
into a certain location. There are two things that make this step
essential.

First, the hash of the archive must match what is expected. Otherwise,
the consumption step fails.

Second, the archive is extracted to an emptyDir volume which is only
accessible by a single Task. This means once the archive is consumed, no
other Tasks can influence its contents. This avoids certain issues with
Time of Check vs Time of Usage, (think parallel tasks sharing a
workspace).

The SLSA Provenance generated for a PipelineRun using these tasks
records the results of the git-clone task, and the parameters for the
buildah task. This allows us to create policies that ensure the buildah
task used the exact archive the git-clone task produced.

I was able to run this in production and produce an image, e.g.
quay.io/redhat-user-workloads/lucarval-tenant/devfile-sample-code-with-quarkus/devfile-sample-code-with-quarkus:3e33c4d702b468aaa46751dc94cf8de5d5233e67

Here's what the SLSA Provenance says for the git-clone task results:

```
$ < /tmp/attestation.json jq '.payload | @base64d | fromjson | .predicate.buildConfig.tasks[] | select(.name == "clone-repository").results[] | select(.name=="source").value.hash'
"6d99e197c4bda4564211943f774561cf85fd0e2ab33226b34eb73d60edaac7aa"
```

And here's what it says for the parameters of the buildah task:

```
$ < /tmp/attestation.json jq '.payload | @base64d | fromjson | .predicate.buildConfig.tasks[] | select(.name == "build-container").invocation.parameters.SOURCE.hash'
"6d99e197c4bda4564211943f774561cf85fd0e2ab33226b34eb73d60edaac7aa"
```

Why bother with this change? This is about ensuring certain segments of
the Pipeline are protected. Today, EC expects all the Tasks on the
Pipeline to be trusted. This is secure, but it's very rigid. It's not
possible for someone to run, for example, a custom Taks to run unit
tests between git-clone and buildah without also customizing the
organization policy. The proposed changes here would allow EC to be more
specific on what it checks instead of just using a general umbrella.

This change is a proof of concept. So it may not fully work as expected.
But hopefully it's enough to help us decide on whether or not to move
forward with this.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>