Emit BASE_IMAGES_DIGESTS result from separate step
Tekton results have a size limit, see
https://tekton.dev/docs/pipelines/tasks/ (search for "termination
messages").

The overall limit for a Task is 4096 bytes, but the limit for each
*step* within a task is lower. And the more steps a task has, the lower
the per-step limit.

The inject-sbom-and-push step emits too many results. The most
problematic one is BASE_IMAGES_DIGESTS, which is a newline-separated
list of the base images for all the stages of an image build.

Move the BASE_IMAGES_DIGESTS to a separate step to reduce the chance of
hitting the size limit.

This is a temporary solution at best, we will need to work out a better
way to pass data between tasks.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>