prefetch: parameterized option to commit changes
When cachi2 updates the Dockerfile this makes the git tree dirty and may
affect scripts or other build processes that use git commit tree
information for things like version numbers and other metadata. When the
git tree has been modified, the version is indicated as being "dirty".

This change allows projects who use this information to inject a `git
commit` in the workflow with the cachi2 changes applied so that and git
version information includes a non-dirty hash.

Note that this does not push that commit back to the origin repository.
I'm not sure if that's problematic or not. It doesn't seem like it
should be.

Signed-off-by: Lance Ball <lball@redhat.com>