Introduce trusted artifacts
Adds the use of trusted artifacts[1] to the Tasks and Pipelines. The
`prefetch-dependencies` Task is no longer skipped based on the
`hermetic` parameter -- it needs to run always to produce the trusted
artifacts as the build* tasks depend on these, so if it were skipped the
build* tasks would be skipped as well.

With this the default value is set for the `*_ARTIFACT` parameters and
the workspace will be unaltered by the `build-trusted-artifact` steps.
This way it is simpler to opt-in to using trusted artifacts, and the
default pipelines do so.

Ref: https://issues.redhat.com/browse/EC-251

[1] https://github.com/redhat-appstudio/build-trusted-artifacts