buildah-rhtap: set the correct Syft catalogers for image scanning
The built image was being mounted as a directory, which makes Syft use a
different set of catalogers than it would use in case it was scanning an
image.

This patch instead saves the image content as an oci-dir, which uses the
same sets of catalogers than scanning an image from a registry or from a
container daemon.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>