Pin Syft's output format to CycloneDX 1.4
Newer versions of Syft may default to newer versions of CycloneDX (as
observed in the recent 0.99.0 version). The problem is that the schema
is bound to change between CycloneDX versions, which can bring breaking
changes to the pipelines, specially to the scripts that merge different
SBOMS.

This patch pins the output CycloneDX format to avoid unwanted breaking
changes.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>