buildah-rhtap: base SBOM component deduplication in the purl
The combination of name+version is not a good enough to guarantee
uniqueness, since there could be cases that components from different
ecosystems (e.g. npm, rpm, golang) share the same name+version.

The purl, on the other hand, takes in consideration not only name and
version, but also the ecosystem, and potentially other things.

There are cases in which the purl is absent, though, namely for
components of type operating-system. Since they are only reported once
in when scanning the image (not the source code), this case alone won't
result in duplication.

We still adding all components that lack a purl (at the risk of
duplication) to avoid errors in scenarios we aren't aware of right now.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>