Enforce YAML Format
This PR requires all files to be formatted with the google YAML formatter before submitting. This can be done with the following commands:

go install github.com/google/yamlfmt/cmd/yamlfmt@v0.10.0 yamlfmt .

This will enforce a standard style that should remove any issues with formatting of PRs.

The actual formatting is done in a subsequent commit.