Always run the `prefetch-dependencies` Task
For trusted artifacts we need the `prefetch-dependencies` Task to run
always. This is because if it is skipped via the `when` conditional the
downstream Tasks that will depend on the trusted artifacts result
produced by `prefetch-dependencies` Task will also be skipped.

This controls the prefetching of the dependencies via the new parameter
`hermetic` which in the default template Pipeline will be passed from
the Pipeline parameter of the same name.