Pin ec-cli image in ec-checks Task
By default OpenShift only fetches an image by a floating tag, e.g.
`snapshot`, if the image is not yet available in the cluster. We can
change this behaver with `imagePullPolicy: Always`. However, a better
approach is to pin the image reference to a digest and let automation,
i.e. renovate bot, bring in updates. This gives us more control over
which code is actually executed in the CI.

I came across this when working on EC-414. The fix for that issue was
not propagated as expected. Further investigation revealed that an older
version of the EC CLI was still being executed.

This commit pins references to the EC CLI image to a digest.

Ref: https://issues.redhat.com/browse/EC-414

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>