buildah-rhtap: Fix SBOM_BLOB_RESULT for images that lack a tag
The algorithm that is being used to strip the tag from the image will
return an empty string in case the image already lacks the tag in the
first place (i.e. quay.io/user/my-image).

This results in a SBOM_BLOB_RESULT that lacks the image name, and only
contains the digest.

Signed-off-by: Bruno Pimentel <bpimente@redhat.com>