Show source image URL in summary
STONEBLD-1608

This is a rework for the reverted version. In this implementation, the
source-build task writes the build result to a file inside workspace so
that any other task can read it. Task summary reads source image URL
from this file and if the result file does not exist yet, just skip
silently.

The workspace mount is optional for task summary. It works for users who
update task summary bundle but not PipelineRun YAML.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>