Store the base image digests in the correct order
Previously the base image digest were not guaranteed to be in the
correct order because the output was generated just from the output of
"buildah images". The output of that command is sorted according to the
creation date of the image, which is not what we want.

The new approach outputs the digests in the order in which they were
found in the dockerfile.

STONEBLD-2203

The "FROM scratch" problem is not yet resolved here. It will be dealt
with in another story/PR as that might be solved on the source-build
task side.

Signed-off-by: mkosiarc <mkosiarc@redhat.com>