Add task that gathers image refs from gitops repo
For RHTAP, we will need a pipeline that runs on pull requests to the
gitops repo. This pipeline will run EC validation on the deployment
images found in the repo.

Add a task that takes a gitops repo (passed in via the 'source'
workspace) and returns the images in EC-compatible format.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>