gitops-pull-request: check only changed images
task/gather-deploy-images:

When the TARGET_BRANCH param is specified, the task will return only the
images that differ between the current revision and the target branch.

Note that the origin/$TARGET_BRANCH reference must already exist in the
cloned repo. The task doesn't have the ability to fetch from private
repos, instead it depends on the clone task to fetch everything
that's necessary.

---

pipelines/gitops-pull-request-rhtap:

* add a target-branch param and pass it to gather-deploy-images
* pass fetchTags: "true" to the clone task (required for
  gather-deploy-images to be able to check the target branch)

Signed-off-by: Adam Cmiel <acmiel@redhat.com>