Save base_images_from_dockerfile to /workspaces/source
The reason is that this was failing in buildah remote task
because the line

echo "$BASE_IMAGES" > /workspace/base_images_from_dockerfile

was executed on a VM via SSH and it was not restored afterwards.
However, /workspaces/source is actually restored via

rsync -ra "$SSH_HOST:$BUILD_DIR/workspaces/source/" "$(workspaces.source.path)/"
https://github.com/redhat-appstudio/build-definitions/blob/main/task/buildah-remote/0.1/buildah-remote.yaml#L326

So storing it in /workspaces/source should make it available in the end.

Signed-off-by: mkosiarc <mkosiarc@redhat.com>