rpm-ostree: use prefetched RPMs when available
STONEBLD-2281

When the prefetch-dependencies task prefetches RPMs, make the
rpm-ostree task use them. Do this by removing all the original repo
files from the source directory and replacing them with the generated
cachi2.repo.

---

Note that the rpm-ostree task determines whether there are prefetched
RPMs by checking the existence of the platform-specific RPMs directory.
This means that if RPMs are prefetched for platform A but not for
platform B, the build for platform B will still work (and download the
RPMs at build time).

One may argue that this is undesirable, as it can lead to a situation
where one platform prefetches RPMs and another one does not. But this is
best addressed by enabling the 'hermetic' option to cut off network
access altogether (to be added in a follow-up commit).

Signed-off-by: Adam Cmiel <acmiel@redhat.com>