Use a newer version of ec in task definition
This should not be merged until *after*
https://issues.redhat.com/browse/EC-518 is resolved, since right now
there are not v0.2 images of ec pushed out to registry.redhat.io.
(Long story.)

It's expected this will resolve the root cause of
https://issues.redhat.com/browse/RHTAPBUGS-1212

I'm aware of the dire warning at the top of this file, but IIUC this
PR to change ref does indeed belong here.