rpm-ostree: add hermetic option
STONEBLD-2281

With HERMETIC=true, the rpm-ostree build will have its network access
disabled.

This is done by setting --network=none for the podman container executed
on the remote VM (which runs the rpm-ostree command). Note that
rpm-ostree has an --offline option, but it doesn't do anything useful
for hermetic builds. Just prints a warning when RPMs have to be
downloaded.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>