buildah: support prefetched RPMs
STONEBLD-2280

When the prefetch task prefetches RPMs for the correct architecture (the
architecture of the machine which runs the buildah task), inject the
generated cachi2.repo into the build.

Use the existing YUM_REPO_D_FETCHED mechanism for this, so that the
cachi2 prefetch can hypothetically be compatible with other prefetch
mechanisms.

Signed-off-by: Adam Cmiel <acmiel1@gmail.com>