rpm-ostree: use prefetched RPMs, add hermetic option (#928)
* prefetch-dependencies: expose dev-package-managers

STONEBLD-2280
STONEBLD-2281

The experimental RPM support that's being added to cachi2 will stay
gated behind the --dev-package-managers flag. Give users a way to enable
dev package managers.

Document the option as discouraged.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>

* rpm-ostree: use prefetched RPMs when available

STONEBLD-2281

When the prefetch-dependencies task prefetches RPMs, make the
rpm-ostree task use them. Do this by removing all the original repo
files from the source directory and replacing them with the generated
cachi2.repo.

---

Note that the rpm-ostree task determines whether there are prefetched
RPMs by checking the existence of the platform-specific RPMs directory.
This means that if RPMs are prefetched for platform A but not for
platform B, the build for platform B will still work (and download the
RPMs at build time).

One may argue that this is undesirable, as it can lead to a situation
where one platform prefetches RPMs and another one does not. But this is
best addressed by enabling the 'hermetic' option to cut off network
access altogether (to be added in a follow-up commit).

Signed-off-by: Adam Cmiel <acmiel@redhat.com>

* rpm-ostree: make podman line more human-friendly

Signed-off-by: Adam Cmiel <acmiel@redhat.com>

* rpm-ostree: add hermetic option

STONEBLD-2281

With HERMETIC=true, the rpm-ostree build will have its network access
disabled.

This is done by setting --network=none for the podman container executed
on the remote VM (which runs the rpm-ostree command). Note that
rpm-ostree has an --offline option, but it doesn't do anything useful
for hermetic builds. Just prints a warning when RPMs have to be
downloaded.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>

---------

Signed-off-by: Adam Cmiel <acmiel@redhat.com>