Fix pull-request pipeline for the GH merge queue
The pipeline used to pass APP_SUFFIX={{ pull_request_number }} to e2e
tests. When a pull request enters the merge queue, it does so via a push
event, not a pull request event. Pipelines as Code therefore does not
expand the {{ pull_request_number }} placeholder, causing e2e tests to
create Kubernetes objects with invalid names:

    "test-app-{{ pull_request_number }}" is invalid: metadata.name: ...

The pull request number in the APP_SUFFIX is also used by e2e test code
to determine the source url and revision for the pull request [1].

To fix the issues:
- use the {{ revision }} instead of the {{ pull_request_number }} for
  the suffix
- pass the source url and revision to the e2e tests directly instead of
  relying on the APP_SUFFIX [2]
- use ref-{{ revision }} as the suffix to make absolutely sure e2e tests
  won't try to interpret the revision as a pull request number

[1]: https://github.com/redhat-appstudio/e2e-tests/blob/c4f5822caebb5b5119eedc4e1ab836eb7e6237f1/tests/build/build_templates.go#L519-L525
[2]: https://github.com/redhat-appstudio/e2e-tests/blob/c4f5822caebb5b5119eedc4e1ab836eb7e6237f1/tests/build/build_templates.go#L534-L537

Signed-off-by: Adam Cmiel <acmiel@redhat.com>