prefetch-dependencies: enable basic git auth
Cachi2 may fetch git tags from the source repository and/or clone other
git repositories. Those may be private. Give the user a way to
authenticate.

The basic-auth mechanism is copy-pasted from the git-clone task, except
that it gets the repository url from 'git remote get-url origin' (unlike
the clone task, which gets it as a Tekton param).

Signed-off-by: Adam Cmiel <acmiel@redhat.com>