Use ga repo for the ec task ref used by RHTAP
** NOTE: As noted mentioned in the comments in the of the file, this change has
no impact on Konflux since Konflux uses the upstream task definition, not this
one. **

Since EC has been declared GA, it's now able to push to the normal repo rather
than the tech preview repo. So let's update the image reference in the
verify-enterprise-contract task accordingly.

Note that this is dependent on
https://issues.redhat.com/browse/EC-570 and
https://issues.redhat.com/browse/EC-510 so I'll keep the PR in draft
until those two are complete.

There's a similar change in the related task definition in the
tssc-sample-pipelines repo.

Ref: https://issues.redhat.com/browse/EC-571