Use ga repo for the ec task ref used by RHTAP
** NOTE: As per the comments in the beginning of the file being
modified, this change has no impact on Konflux since Konflux uses
the upstream task definition, not this one. **

Since EC has been declared GA, it's now able to push to the normal
repo rather than the tech preview repo. So let's update the image
reference in the verify-enterprise-contract task to use the EC
image in the new GA repo.

There's a similar change in the related task definition in the
tssc-sample-pipelines repo, see
https://github.com/redhat-appstudio/tssc-sample-pipelines/pull/39 .

Ref: https://issues.redhat.com/browse/EC-571